requirements.txt
numpy==1.20.1
pandas==1.2.3
pyarrow==3.0.0
python-dateutil==2.8.1
requests==2.25.1
